+++ src/nougaro.py
@@ -113,7 +113,8 @@ def run(
         work_dir = noug_dir
     interpreter = src.runtime.interpreter.Interpreter(run, noug_dir, new_args_strings, work_dir)
     if use_context is None:
-        context = Context('<program>')  # create the context of the interpreter
+        assert tokens[0].pos_start is not None
+        context = Context('<program>', tokens[0].pos_start, None)  # create the context of the interpreter
         # don't forget to change the context symbol table to the global symbol table
         if use_default_symbol_table:
             context.symbol_table = default_symbol_table.copy()